Calliope-Rennspiel/Python/requirements.txt
pyserial==3.5
pygame==2.1.3
scikit-learn==1.1.3
pandas==1.5.3
pynput==1.7.6
orange3==3.34.1
serial-tool==0.0.1
wheel==0.41.2
